requirements.txt
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
sentence-transformers>=2.2.0
transformers>=4.30.0
peft>=0.4.0
pillow>=10.0.0
scipy>=1.10.0
